New design
New design proposal! Everything isn't in yet (like sign in and publish
all*)

* I added a publish button to each message so you can publish only that
message if you want to. Also a "Publish all" button is going to be
added.

In this version a new message is added to the main view when editing
and when done sent to edit view. This is going to be changed, this was
just the fastest way (for me) to test the new design.